requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
langgraph==0.2.55
langgraph-checkpoint-sqlite==2.0.4
langchain==0.3.9
langchain-groq==0.2.1
langchain-openai==0.3.1
langchain-community==0.3.9
python-dotenv==1.0.1
pydantic==2.10.2
httpx==0.27.2
aiosqlite==0.20.0
duckduckgo-search==6.3.3
sse-starlette==2.1.3
python-multipart==0.0.17
